Use dotenv to configure server environment
This will allow the environment variables for the instance on the
server to be stored in an ".env" file. For example, the production
environment might have a value set for the "SECRET" environemnt
variable and a test instance, which may reside on the same machine,
can have a different value.